backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
Flask-Mail==0.9.1
werkzeug==2.3.7